+++ test/controller/error/internal.js
@@ -1,3 +1,9 @@
 module.exports.GET = function (req, res) {
 	res.json('internal error');
 };
+module.exports.POST = function (req, res) {
+	res.json('internal error');
+};
+module.exports.PUT = function (req, res) {
+	res.json('internal error');
+};