+++ test/controller/error/notFound.js
@@ -1,3 +1,9 @@
 module.exports.GET = function (req, res) {
 	res.html('not found');
 };
+module.exports.POST = function (req, res) {
+	res.html('not found');
+};
+module.exports.PUT = function (req, res) {
+	res.html('not found');
+};